Crate near_vm_errors

Modules

hex_format

Enums

CacheError

CompilationError

EvmError
Errors specifically from native EVM.

FunctionCallError

HostError

InconsistentStateError
An error that is caused by an operation on an inconsistent state. E.g. a deserialization error or an integer overflow.

MethodResolveError

PrepareError
Error that can occur while preparing or executing Wasm smart-contract.

VMError

VMLogicError

WasmTrap
A kind of a trap happened during execution of a binary